% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reference.R
\name{add_reference}
\alias{add_reference}
\title{Add a BibEntry Reference to an Existing Collection}
\usage{
add_reference(bib, file)
}
\arguments{
\item{bib}{One more \code{BibEntry}-class objects as defined by \code{RefManageR::is.BibEntry()}}

\item{file}{Name (as a \code{character} string) of a .bib file containing \code{BibEntry}-class objects}
}
\description{
Merge one or more BibEntries together without 
             duplicating entries
}
\details{
stuff
}
\examples{
\dontrun{
require(RefManageR)

# Create First BibEntry
ref <- as.BibEntry(citation(package = 'RefManageR'))

# Write BibEntry to file new_bib.bib
WriteBib(bib = ref, file = 'new_bib.bib')

# Create Second BibEntry
surv <- as.BibEntry(citation(package = 'survival'))

# Add second BibEntry to File
add_reference(bib = surv, file = 'new_bib.bib')
}
}
