% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afit_far.R
\name{afit_far}
\alias{afit_far}
\title{Air Force Institute of Technology Thesis Template}
\usage{
afit_far(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE, toc = TRUE, template = "template.tex",
  format = "afit_far")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{fig_caption}{Should figure captions be enabled?}

\item{toc}{Should a table of contents be created?}

\item{template}{The TeX template used}

\item{format}{The format of the document}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating an annual AFIT Faculty Activity Report
}
\details{
Adapted from the STYLE GUIDE FOR AFIT DISSERTATIONS, THESES AND GRADUATE RESEARCH PAPERS - July 2015
}
\examples{

\dontrun{
rmarkdown::draft("MyThesis.Rmd", template = "afit_thesis", package = "AFIT")
}

}
\seealso{
\code{\link{style_guide(school = 'afit')}}
}
