% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afit_presentation.R
\name{afit_presentation}
\alias{afit_presentation}
\title{Create an AFIT Presentation}
\usage{
afit_presentation(
  ...,
  logo = NULL,
  template = "default.html",
  format = "afit_presentation",
  incremental = FALSE,
  fig_width = 8,
  fig_height = 4.9,
  self_contained = F,
  smart = FALSE,
  css = NULL,
  js = NULL,
  highlight = "default",
  pandoc_args = NULL,
  duration = NULL,
  footer = NULL,
  includes = NULL,
  mathjax = "default",
  font_adjustment = 0
)
}
\arguments{
\item{...}{extra stuff}

\item{logo}{The file path for a image file shown on the title slide (relative to the .Rmd file)}

\item{template}{The HTML template file}

\item{format}{The type of template format type}

\item{incremental}{Should bullets be revealed individually?}

\item{fig_width}{The default figure width}

\item{fig_height}{The default figure height}

\item{self_contained}{How to handle resources}

\item{smart}{Is this presentation smart}

\item{css}{The file path for a css file (relative to the .Rmd file)}

\item{js}{The file path for a js file (relative to the .Rmd file)}

\item{highlight}{Code highlighting applied}

\item{pandoc_args}{Arguments to apply to Pandoc}

\item{duration}{Duration (in minutes) of the slide deck. This value is used
to add a countdown timer to the slide footer.}

\item{footer}{Footer text (e.g. organization name and/or copyright)}

\item{includes}{Additional content to include using \code{rmarkdown::includes}}

\item{mathjax}{Should mathjax be included}

\item{font_adjustment}{Increase or decrease the default font size
(e.g. -1 or +1). You can also manually adjust the font size during the
presentation using the 'S' (smaller) and 'B' (bigger) keys.}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a slidy presentation.
}
\details{
See the \href{http://rmarkdown.rstudio.com/slidy_presentation_format.html}{online
documentation} for additional details on using the \code{slidy_presentation} format.

For more information on markdown syntax for presentations see the
\href{http://pandoc.org/README.html}{pandoc online documentation}.
}
\examples{
\dontrun{

library(AFIT)

# simple invocation
render("afit_presentation.Rmd", slidy_presentation())

# specify an option for incremental rendering
render("afit_presentation.Rmd", slidy_presentation(incremental = TRUE))
}

}
