% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afit_thesis.R
\name{afit_thesis}
\alias{afit_thesis}
\title{Air Force Institute of Technology Thesis Template}
\source{
STYLE GUIDE FOR AFIT DISSERTATIONS, THESES AND GRADUATE RESEARCH PAPERS - July 2015
}
\usage{
afit_thesis(..., keep_tex = !FALSE,
  md_extensions = c("-autolink_bare_uris"), fig_caption = TRUE,
  toc = TRUE, template = "templateb.tex", format = "afit_thesis",
  highlight = "default", xtab_caption_placement = "top",
  xtab_comment = FALSE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{fig_caption}{Should figure captions be enabled?}

\item{toc}{Should a table of contents be created?}

\item{template}{The TeX template used}

\item{format}{The format of the document}

\item{highlight}{Highlight color template for code chunks}

\item{xtab_caption_placement}{Placement of captions for tables created using the \code{xtable} package}

\item{xtab_comment}{Logical variable whether the \code{xtable} timestamp should be printed}
}
\value{
R Markdown output format to pass to \code{rmarkdown::render}
}
\description{
Create an AFIT thesis using rmarkdown
}
\examples{

\dontrun{
rmarkdown::draft("MyThesis.Rmd", template = "afit_thesis", package = "DODschools")
}

}
\seealso{
\code{\link{style_guide}}
}
