% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{build_forms}
\alias{build_forms}
\title{Build One or More Administrative Forms}
\usage{
build_forms(yml = "metadata.yml", which = NULL, ...)
}
\arguments{
\item{yml}{\code{character} Path to the metadata.yml file}

\item{which}{A \code{numeric} ID corresponding to the form to be built. If \code{NULL} (default), all forms are built. (see details)}

\item{...}{Currently not used}
}
\description{
Automatically create and fill forms required for 
             graduation using specified metadata
}
\details{
This function should only be used after a template directory has been created.
         When invoked, this function reads the YAML metadata contained within 
         the file provided as the argument \code{yml} and passes these variables 
         to the LaTeX form template specified by \code{which}. 
         
         The \code{which} argument accepts integer values from one (1) to five (5),
         where the values specify forms as
         
         \itemize{
           \item{'1'}{'Standard Form 298 (SF298)'}
           \item{'2'}{'Distribution Statement Form'}
           \item{'3'}{'Public Affairs Approval Request Form'}
           \item{'4'}{'Committee Signature Form'}
           \item{'5'}{'AF Form 475 - Training Report'}
         }

         NOTE: All of the templates used to render these forms utilize the 
         eforms LaTeX package.  As result each form can be edited after 
         rendering and can also be electronically signed.
}
