% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_module.R
\name{canvas_module}
\alias{canvas_module}
\title{Create a Module that conforms to the Instructure Canvas API}
\usage{
canvas_module(
  ...,
  keep_md = !FALSE,
  md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE,
  toc = TRUE,
  highlight = "default"
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::html_document}}

\item{keep_md}{Should the raw md files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{fig_caption}{Should figure captions be enabled?}

\item{toc}{Should a table of contents be created?}

\item{highlight}{Highlight color template for code chunks}

\item{template}{The TeX template used}

\item{format}{The format of the document}
}
\value{
R Markdown output format to pass to \code{rmarkdown::render}
}
\description{
Create an Canvas Module using rmarkdown
}
\examples{

\dontrun{
rmarkdown::draft("MyModule.Rmd", template = "canvas_module", package = "DODschools")
}

}
