% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeDTIC.R
\name{citeDTIC}
\alias{citeDTIC}
\title{Create a BibEntry from an international standard book number (ISBN)}
\source{
http://www.dtic.mil/dtic/
}
\usage{
citeDTIC(designator = "AFIT-ENS-DS-13-D-02", access_num = NULL, ...)
}
\arguments{
\item{designator}{An AFIT thesis or dissertation designator number}

\item{access_num}{The DTIC access number}

\item{...}{extra stuff}
}
\description{
This function uses advanced search on dtic.mil to generate a bibtex entry for books using the 
ISBN.  The result is then formatted into a bibentry-class object using \code{knitcitations::greycite}and
}
\details{
\code{isbn} may contain dashes
}
