% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeISBN.R
\name{citeISBN}
\alias{citeISBN}
\title{Create a BibEntry from an international standard book number (ISBN)}
\source{
http://www.ottobib.com/
}
\usage{
citeISBN(
  isbn = NULL,
  preview = TRUE,
  bibtype = NULL,
  key = NULL,
  author = NULL,
  title = NULL,
  publisher = NULL,
  year = NULL,
  address = NULL,
  style = "bibtex",
  ...
)
}
\arguments{
\item{isbn}{An international standard book number (ISBN-13 or ISBN-10)}

\item{preview}{If \code{TRUE}, the bibentry is printed to the screen via \code{cat}}

\item{bibtype}{The type of bibliographic element. (see \code{details})}

\item{key}{The bibtex key for the bibentry}

\item{author}{The author for the bibentry}

\item{title}{The title for the bibentry}

\item{publisher}{The publisher for the bibentry}

\item{year}{The year for the bibentry}

\item{address}{The publisher's address for the bibentry}

\item{style}{The output style of the bibentry (defaults to 'bibtex')}

\item{...}{extra stuff}
}
\description{
This function uses ottobib.com to generate a bibtex entry for books using the 
ISBN.  The result is then formatted into a bibentry-class object using \code{RegManageR::BibEntry}and
}
\details{
\code{isbn} may contain dashes
}
