% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{distro_form}
\alias{distro_form}
\title{Air Force Document Distribution Statement Form}
\usage{
distro_form(
  ...,
  keep_tex = FALSE,
  md_extensions = c("-autolink_bare_uris"),
  template = "distroForm.tex"
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{template}{The TeX template used}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Air Force Document Distribution Statement Form
}
