% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DODschools-package.R
\docType{package}
\name{DODschools-package}
\alias{DODschools}
\alias{DODschools-package}
\title{DODschools: Academic Document Templates for Department of Defense Schools}
\description{
DODschools: Academic Document Templates for Department of Defense Schools
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auburngrads/DODschools}
  \item Report bugs at \url{https://github.com/Auburngrads/DODschools/issues}
}

}
\author{
\strong{Maintainer}: Jason Freels \email{auburngrads@live.edu}

Other contributors:
\itemize{
  \item Corey Degroot \email{corey.degroot@us.af.mil} [contributor]
  \item Diane Schultz \email{diane.schultz@us.af.mil} [contributor]
  \item Joseph Buyer \email{joseph.buyer@us.af.mil} [contributor]
  \item James Daniels \email{jim.daniels@us.af.mil} [contributor]
  \item Christopher Twigg \email{christopher.twigg@us.af.mil} [contributor]
}

}
