% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dontTouch.R
\name{dontTouch}
\alias{dontTouch}
\title{Extract YAML data from metadata.yml}
\usage{
dontTouch(file = NULL)
}
\arguments{
\item{file}{Path to the YAML file}
}
\description{
Extract YAML data from metadata.yml
}
