% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file}
\alias{find_file}
\title{Helper function to find a file for a named template}
\usage{
find_file(template, file)
}
\arguments{
\item{template}{A named template folder}

\item{file}{the name of the file to search for within the \code{template} folder}
}
\description{
Helper function to find a file for a named template
}
