% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_resource}
\alias{find_resource}
\title{Helper function to find a resource for a named template}
\usage{
find_resource(template, file)
}
\arguments{
\item{template}{A named template folder}

\item{file}{Name of the  resource file to search for within the \code{template} folder}
}
\description{
Helper function to find a resource for a named template
}
