% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYAML.R
\name{getYAML}
\alias{getYAML}
\title{Extract YAML metadata from an Rmarkdown file}
\usage{
getYAML(file)
}
\arguments{
\item{file}{A file containing a YAML metadata header}
}
\description{
Extract YAML metadata from an Rmarkdown file
}
