% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieee_article.R
\name{ieee_article}
\alias{ieee_article}
\title{Air Force Institute of Technology Thesis Template}
\usage{
ieee_article(
  ...,
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE,
  toc = TRUE,
  template = "templateb.tex",
  format = "ieee_article"
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{fig_caption}{Should figure captions be enabled?}

\item{toc}{Should a table of contents be created?}

\item{template}{The TeX template used}

\item{format}{The format of the document}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating AFIT theses
}
\details{
Adapted from the STYLE GUIDE FOR AFIT DISSERTATIONS, THESES AND GRADUATE RESEARCH PAPERS - July 2015
}
\examples{

\dontrun{
rmarkdown::draft("MyThesis.Rmd", template = "afit_thesis", package = "AFIT")
}

}
