% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inherit_html_document}
\alias{inherit_html_document}
\title{Inherits properties from an existing html document}
\usage{
inherit_html_document(...)
}
\arguments{
\item{...}{arguments sent to \code{rmarkdown::html_document}}
}
\description{
Calls rmarkdown::html_document and marks the return value as inheriting html_document
}
