% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inherit_pdf_document}
\alias{inherit_pdf_document}
\title{Inherits properties from an existing pdf document}
\usage{
inherit_pdf_document(...)
}
\arguments{
\item{...}{arguments sent to \code{rmarkdown::pdf_document}}
}
\description{
Calls rmarkdown::pdf_document and marks the return value as inheriting pdf_document
}
