% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noTouch.R
\name{noTouch}
\alias{noTouch}
\title{Extract YAML metadata from metadata.yml}
\usage{
noTouch(file = NULL)
}
\arguments{
\item{file}{Path to the metadata.yaml file}
}
\description{
Extract YAML metadata from metadata.yml
}
