% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf_document_format}
\alias{pdf_document_format}
\title{Create a custom format derived from pdf_document}
\usage{
pdf_document_format(..., format, template, csl = NULL)
}
\arguments{
\item{...}{arguments sent to \code{inherit_pdf_document}}

\item{format}{Named format folder}

\item{template}{Named template folder}

\item{csl}{Named style file for a given \code{template}}
}
\description{
Helper function to create a custom format derived from pdf_document that includes a custom LaTeX template and custom CSL definition
}
