% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{public_affairs}
\alias{public_affairs}
\title{Air Force Public Affairs Review Form}
\usage{
public_affairs(
  ...,
  keep_tex = FALSE,
  md_extensions = c("-autolink_bare_uris"),
  fig_caption = FALSE,
  toc = FALSE,
  template = "PAform.tex"
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{fig_caption}{Should figure captions be enabled?}

\item{toc}{Should a table of contents be created?}

\item{template}{The TeX template used}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Air Force Public Affairs Review Form
}
