% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run examples in packages}
\usage{
runExample(pkg, preview.files = TRUE, which = NULL, output = "html",
  shiny = FALSE)
}
\arguments{
\item{pkg}{[\code{character}] An installed package}

\item{preview.files}{[\code{logical}] If \code{TRUE}, a numbered list of available examples is returned}

\item{which}{[\code{numerical}] Which example should be run?}

\item{output}{[\code{character}] File extension of the output file}

\item{shiny}{[\code{logical}] If example contains shiny content, \code{TRUE}}
}
\description{
This function retreives examples stored in the Examples folder for installed packages.
}
