% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{sf298}
\alias{sf298}
\title{Air Force Form SF 298}
\usage{
sf298(
  ...,
  keep_tex = FALSE,
  md_extensions = c("-autolink_bare_uris"),
  template = "sf298.tex"
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{template}{The TeX template used}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Air Force Form SF 298
}
