% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{training_report}
\alias{training_report}
\title{Air Force Training Report}
\usage{
training_report(..., keep_tex = FALSE,
  md_extensions = c("-autolink_bare_uris"), template = "trainingReport.tex")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{template}{The TeX template used}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Air Force Training Report
}
