% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{buildForms}
\alias{buildForms}
\title{Build one or more forms}
\usage{
buildForms(yml = "metadata.yml", which = NULL, ...)
}
\arguments{
\item{yml}{\code{character} Path to the metadata.yml file}

\item{which}{\code{numeric} Which form to build. If \code{NULL} all forms are built.}
}
\description{
Build one or more forms
}
