% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeDTIC.R
\name{citeDTIC}
\alias{citeDTIC}
\title{Create a BibEntry from a DTIC document (Experimental)}
\source{
http://www.dtic.mil/dtic/
}
\usage{
citeDTIC(designator = "AFIT-ENS-DS-13-D-02", access_num = NULL, ...)
}
\arguments{
\item{designator}{An AFIT thesis or dissertation designator number}

\item{access_num}{The DTIC access number}

\item{...}{extra stuff}
}
\description{
Generate reference from document stored within the DTIC repository
}
\details{
This function references the advanced search utility on dtic.mil 
         to generate a bibtex.  Results returned from the search are formatted 
         as a bibentry-class object using \code{knitcitations::greycite}. 
         
         Note: the \code{designator} argument may contain dashes.
}
