% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DODschools-package.R
\docType{package}
\name{DODschools-package}
\alias{DODschools}
\alias{DODschools-package}
\title{DODschools: Academic Document Templates for Department of Defense Schools}
\description{
Rmarkdown document and presentation templates comforming to the style guides for 
    Department of Defense (DoD) academic institutions. As of Verison 1.0.0, templates 
    have been developed for the DoD graduate schools: the Air Force Institute of Technology 
    (AFIT) and the Naval Postgraduate School (NPS).  Subsequent versions will include 
    templates for the United States Military Academy (USMA), the United States Naval Academy 
    (USNA), and the United States Air Force Academy (USAFA).  Includes functions to 
    auto-generate required forms and import reference metadata from works published to the 
    Defense Technical Information Center (DTIC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auburngrads/DODschools}
  \item Report bugs at \url{https://github.com/Auburngrads/DODschools/issues}
}

}
\author{
\strong{Maintainer}: Jason Freels \email{auburngrads@live.edu}

Other contributors:
\itemize{
  \item Corey Degroot \email{corey.degroot@us.af.mil} [contributor]
  \item Diane Schultz \email{diane.schultz@us.af.mil} [contributor]
  \item Joseph Buyer \email{joseph.buyer@us.af.mil} [contributor]
  \item James Daniels \email{jim.daniels@us.af.mil} [contributor]
  \item Christopher Twigg \email{christopher.twigg@us.af.mil} [contributor]
}

}
