% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{sf298}
\alias{sf298}
\title{Air Force Institute of Technology Prospectus Template}
\usage{
sf298(..., keep_tex = FALSE, md_extensions = c("-autolink_bare_uris"),
  template = "sf298.tex")
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{template}{The TeX template used}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating the initial pages of AFIT theses, dissertations, and reports.
}
\details{
Adapted from the STYLE GUIDE FOR AFIT DISSERTATIONS, THESES AND GRADUATE RESEARCH PAPERS - July 2015
\href{http://www.afit.edu/docs/NewAFITStyleGuideJuly2015(2).pdf}{http://www.afit.edu/docs/NewAFITStyleGuideJuly2015(2).pdf}
}
\examples{

\dontrun{
rmarkdown::draft("MyThesis.Rmd", template = "afit_prospectus", package = "AFIT")
}

}
